ProactiveChatController converted to TS.
